# internal/server/static/openapi.json
{
    "swagger": "2.0",
    "info": {
        "description": "✨Reverse-engineered API for Gemini web app. It can be used as a genuine API key from OpenAI, Gemini, and Claude.",
        "title": "Gemini Web To API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4981",
    "basePath": "/",
    "paths": {
...
        "/gemini/v1beta/deepresearch": {
            "post": {
                "description": "Performs deep research on a topic using Gemini.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gemini"
                ],
                "summary": "Deep Research (synchronous)",
                "parameters": [
                    {
                        "description": "Deep Research Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.DeepResearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.DeepResearchResponse"
                        }
                    }
                }
            }
        },
...
        "/gemini/v1beta/models/{model}:generateContent": {
            "post": {
                "description": "Generates content using the Gemini model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gemini"
                ],
                "summary": "Generate Content (Gemini)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "model",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Generate Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.GeminiGenerateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.GeminiGenerateResponse"
                        }
                    }
                }
            }
        },
        "/gemini/v1beta/models/{model}:streamGenerateContent": {
            "post": {
                "description": "Streams generated content using the Gemini model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gemini"
                ],
                "summary": "Stream Generate Content (Gemini)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "model",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Generate Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.GeminiGenerateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chunked JSON response",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gemini-web-to-api_internal_modules_gemini_dto.Candidate": {
            "type": "object",
            "properties": {
                "content": {
                    "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.Content"
                },
                "finishMessage": {
                    "type": "string"
                },
                "finishReason": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.Content": {
            "type": "object",
            "properties": {
                "parts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.Part"
                    }
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.DeepResearchRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "description": "Language for the response (optional, e.g. \"en\", \"vi\")",
                    "type": "string"
                },
                "max_sources": {
                    "description": "MaxSources maximum number of web sources to consult (optional)",
                    "type": "integer"
                },
                "model": {
                    "description": "Model to use (optional, defaults to best available)",
                    "type": "string"
                },
                "query": {
                    "description": "Query is the research topic or question",
                    "type": "string"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.DeepResearchResponse": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "CompletedAt timestamp (Unix seconds), 0 if not completed",
                    "type": "integer"
                },
                "created_at": {
                    "description": "CreatedAt timestamp (Unix seconds)",
                    "type": "integer"
                },
                "duration_ms": {
                    "description": "DurationMs total research duration in milliseconds",
                    "type": "integer"
                },
                "error": {
                    "description": "Error message if status is \"failed\"",
                    "type": "string"
                },
                "id": {
                    "description": "ID of this research session",
                    "type": "string"
                },
                "model": {
                    "description": "Model used",
                    "type": "string"
                },
                "query": {
                    "description": "Query that was researched",
                    "type": "string"
                },
                "sources": {
                    "description": "Sources used in research",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.ResearchSource"
                    }
                },
                "status": {
                    "description": "Status: \"in_progress\", \"completed\", \"failed\"",
                    "type": "string"
                },
                "steps": {
                    "description": "Steps of research performed",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.ResearchStep"
                    }
                },
                "summary": {
                    "description": "Summary is the synthesized research report",
                    "type": "string"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.FunctionCall": {
            "type": "object",
            "properties": {
                "args": {
                    "description": "@SchemaType object",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.FunctionCallingConfig": {
            "type": "object",
            "properties": {
                "allowed_function_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mode": {
                    "description": "\"AUTO\", \"ANY\", \"NONE\"",
                    "type": "string"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.FunctionDeclaration": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parameters": {
                    "description": "@SchemaType object",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.FunctionResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "response": {
                    "description": "@SchemaType object",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.GeminiGenerateRequest": {
            "type": "object",
            "properties": {
                "contents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.Content"
                    }
                },
                "generationConfig": {
                    "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.GenerationConfig"
                },
                "safety_settings": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                },
                "tool_config": {
                    "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.ToolConfig"
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.Tool"
                    }
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.GeminiGenerateResponse": {
            "type": "object",
            "properties": {
                "candidates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.Candidate"
                    }
                },
                "usageMetadata": {
                    "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.UsageMetadata"
                }
            }
        },
...
        "gemini-web-to-api_internal_modules_gemini_dto.GenerationConfig": {
            "type": "object",
            "properties": {
                "maxOutputTokens": {
                    "type": "integer"
                },
                "temperature": {
                    "type": "number"
                },
                "topK": {
                    "type": "integer"
                },
                "topP": {
                    "type": "number"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.InlineData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.Part": {
            "type": "object",
            "properties": {
                "function_call": {
                    "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.FunctionCall"
                },
                "function_response": {
                    "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.FunctionResponse"
                },
                "inline_data": {
                    "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.InlineData"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.ResearchSource": {
            "type": "object",
            "properties": {
                "domain": {
                    "description": "Domain of the source",
                    "type": "string"
                },
                "snippet": {
                    "description": "Snippet relevant excerpt from the source",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the source page",
                    "type": "string"
                },
                "url": {
                    "description": "URL of the source",
                    "type": "string"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.ResearchStep": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description of what this step did",
                    "type": "string"
                },
                "query": {
                    "description": "Query used in this step (for search steps)",
                    "type": "string"
                },
                "result": {
                    "description": "Result summary of this step",
                    "type": "string"
                },
                "step_number": {
                    "description": "StepNumber ordinal step number",
                    "type": "integer"
                },
                "type": {
                    "description": "Type of step: \"search\", \"read\", \"synthesize\", \"outline\", \"write\"",
                    "type": "string"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.Tool": {
            "type": "object",
            "properties": {
                "function_declarations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.FunctionDeclaration"
                    }
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.ToolConfig": {
            "type": "object",
            "properties": {
                "function_calling_config": {
                    "$ref": "#/definitions/gemini-web-to-api_internal_modules_gemini_dto.FunctionCallingConfig"
                }
            }
        },
        "gemini-web-to-api_internal_modules_gemini_dto.UsageMetadata": {
            "type": "object",
            "properties": {
                "candidatesTokenCount": {
                    "type": "integer"
                },
                "promptTokenCount": {
                    "type": "integer"
                },
                "totalTokenCount": {
                    "type": "integer"
                }
            }
        },
...
    }
}